long long is long
git-svn-id: svn://hinux.hin.no/home/DT2006/murray/svn/gh@213 c6f41995-d01f-4a11-924a-98aa25776321

+++ trunk/grouphug/modules/WordCount.java
@@ -35,7 +35,7 @@ public class WordCount implements GrouphugModule {
 				sql.query("INSERT INTO "+WORDS_DB+" (nick, words) VALUES ('"+sender+"', '"+count+"');");
 			}else{
 				Object[] values = sql.getValueList();
-				sql.query("UPDATE "+WORDS_DB+" SET count='"+((Integer)values[2] + count)+"' WHERE id='"+values[0]+"';");
+				sql.query("UPDATE "+WORDS_DB+" SET count='"+((Long)values[2] + count)+"' WHERE id='"+values[0]+"';");
 			}
 
 		}catch(SQLException e) {
